requirements.txt
absl-py==1.0.0
accelerate==0.34.2
annotated-types==0.7.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
chart-studio==1.1.0
cloudpickle==3.0.0
coloredlogs==15.0.1
contextlib2==21.6.0
cppimport==22.8.2
cycler==0.11.0
dicom2nifti==2.3.4
diffusers==0.30.3
einops==0.4.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.31.1
fsspec==2024.9.0
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.2
Jinja2==3.1.4
joblib==1.1.0
kiwisolver==1.4.0
linecache2==1.0.0
Markdown==3.3.6
markdown-it-py==3.0.0
matplotlib==3.5.1
mdurl==0.1.2
MedPy==0.4.0
ml_collections==0.1.1
monai==1.3.2
mpmath==1.3.0
netron==7.8.9
networkx==2.6.3
nibabel==3.2.2
ninja==1.11.1.1
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-modelopt==0.17.0
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
onnx==1.16.1
onnx-graphsurgeon==0.5.2
onnxconverter-common==1.14.0
onnxmltools==1.12.0
onnxruntime-gpu==1.18.0
onnxruntime_extensions==0.12.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.5
patsy==0.5.2
Pillow==9.0.1
plotly==5.6.0
polygraphy==0.49.9
protobuf==3.20.2
psutil==6.0.0
ptflops==0.6.8
PuLP==2.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.13.6
pydantic==2.9.2
pydantic_core==2.23.4
pydicom==2.3.0
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
regex==2024.9.11
requests==2.27.1
requests-oauthlib==1.3.1
retrying==1.3.3
rich==13.8.1
rsa==4.8
safetensors==0.4.5
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.10.1
seaborn==0.11.2
SimpleITK
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
sympy==1.13.3
tenacity==8.0.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.5.4
tokenizers==0.19.1
torch
torchprofile==0.0.4
torchvision==0.19.1
tqdm==4.63.0
traceback2==1.4.0
transformers==4.44.2
triton==3.0.0
unittest2==1.1.0
urllib3==1.26.8
webcolors==1.11.1
Werkzeug==2.0.3